requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0